requirements.txt
Django==5.0.1
django_environ==0.11.2
django_shortuuidfield==0.1.3
psycopg2==2.9.9
Requests==2.31.0
